Ajustes em banco de dados, mensagens de retorno e feito validação de permissões para ações em projetos e tarefas

+++ app/Modules/Project/Application/Services/ProjectService.php
@@ -2,13 +2,18 @@
 
 namespace App\Modules\Project\Application\Services;
 
+use App\Modules\Project\Application\Events\ProjectCreated;
 use App\Modules\Project\Domain\Entities\Project;
 use App\Modules\Project\Domain\Repositories\ProjectRepositoryInterface;
+use App\Modules\Shared\Authorization\PermissionService;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Support\Facades\Event;
 
 class ProjectService
 {
     public function __construct(
-        protected ProjectRepositoryInterface $repo
+        protected ProjectRepositoryInterface $repo,
+        protected PermissionService $permissionService,
     ) {}
 
     public function list(int $userId): array
@@ -16,9 +21,14 @@ class ProjectService
         return $this->repo->findAllByUser($userId);
     }
 
-    public function create(string $name, int $ownerId): Project
+    public function create(array $data, int $userId): Project
     {
-        $project = new Project(id: 0, name: $name, ownerId: $ownerId);
+        $project = new Project(
+            id: 0,
+            name: $data['name'],
+            ownerId: $userId,
+            description: $data['description']
+        );
         return $this->repo->save($project);
     }
 
@@ -27,14 +37,34 @@ class ProjectService
         return $this->repo->findById($id);
     }
 
-    public function update(int $id, string $name, int $ownerId): ?Project
+    public function update(int $id, array $data, int $userId): ?Project
     {
-        $project = new Project(id: $id, name: $name, ownerId: $ownerId);
-        return $this->repo->save($project);
+        if (!$this->permissionService->canEditProject($userId, $id)) {
+            throw new AuthorizationException('Você não tem permissão para editar este projeto.');
+        }
+
+        $project = $this->repo->findById($id);
+        if (!$project) {
+            throw new \Exception("Projeto não encontrado!");
+        }
+
+        $project->name = $data['name'];
+        $project->description = $data['description'] ?? '';
+        $project->ownerId = $data['owner_id'] ?? $userId;
+
+        $savedProject = $this->repo->save($project);
+
+        Event::dispatch(new ProjectCreated($savedProject->toArray()));
+
+        return $savedProject;
     }
 
-    public function delete(int $id): void
+    public function delete(int $id, int $userId): void
     {
+        if (!$this->permissionService->canDeleteProject($userId, $id)) {
+            throw new AuthorizationException('Você não tem permissão para deletar este projeto.');
+        }
+
         $this->repo->delete($id);
     }
 }
